LineHotel/requirements.txt
flask
line-bot-sdk
firebase-admin
gunicorn
